Sparrow dom/metastrategy m 01 tests (#1153)
* operate with exepcted token balances instead of amounts that contract holds

* prettier

* slither

* remove metapoolLP tokens contract balance from checkBalance calculation

* do not pull funds from compound dai. Messes up vault valuations

* correct minimum amount calculation and fix tests

* minor fix

* tests

* cleaner

* test fixes

* prettier

* test name fixes

* improve fork test speed

* improve the resiliance of the tests

* prettier

* fix tests
