+++ contracts/test/_fixture.js
@@ -1063,4 +1063,5 @@ module.exports = {
   hackedVaultFixture,
   rebornFixture,
   withImpersonatedAccount,
+  impersonateAndFundContract,
 };
