+++ contracts/test/strategies/ousd-metapool-3crv-tilted-pool.fork-test.js
@@ -1,31 +1,31 @@
 const { expect } = require("chai");
 
 const { loadFixture } = require("ethereum-waffle");
-const { units, ousdUnits, forkOnlyDescribe } = require("../../helpers");
-const { withBalancedOUSDMetaPool } = require("../../_metastrategies-fixtures");
+const { units, ousdUnits, forkOnlyDescribe } = require("../helpers");
+const { withCRV3TitledOUSDMetapool } = require("../_metastrategies-fixtures");
 
 forkOnlyDescribe(
-  "ForkTest: Convex 3pool/OUSD Meta Strategy - Balanced Metapool",
+  "ForkTest: Convex 3pool/OUSD Meta Strategy - Titled to 3CRV",
   function () {
     this.timeout(0);
     // due to hardhat forked mode timeouts - retry failed tests up to 3 times
     this.retries(3);
 
     describe("Mint", function () {
-      it("Should stake USDT in Cruve guage via metapool", async function () {
-        const fixture = await loadFixture(withBalancedOUSDMetaPool);
+      it("Should stake USDT in Curve guage via metapool", async function () {
+        const fixture = await loadFixture(withCRV3TitledOUSDMetapool);
         const { josh, usdt } = fixture;
-        await mintTest(fixture, josh, usdt, "100000");
+        await mintTest(fixture, josh, usdt, "200000");
       });
 
-      it("Should stake USDC in Cruve guage via metapool", async function () {
-        const fixture = await loadFixture(withBalancedOUSDMetaPool);
+      it("Should stake USDC in Curve guage via metapool", async function () {
+        const fixture = await loadFixture(withCRV3TitledOUSDMetapool);
         const { matt, usdc } = fixture;
-        await mintTest(fixture, matt, usdc, "120000");
+        await mintTest(fixture, matt, usdc, "110000");
       });
 
-      it("Should NOT stake DAI in Cruve guage via metapool", async function () {
-        const fixture = await loadFixture(withBalancedOUSDMetaPool);
+      it("Should NOT stake DAI in Curve guage via metapool", async function () {
+        const fixture = await loadFixture(withCRV3TitledOUSDMetapool);
         const { anna, dai } = fixture;
         await mintTest(fixture, anna, dai, "110000");
       });
@@ -34,7 +34,7 @@ forkOnlyDescribe(
     describe("Redeem", function () {
       it("Should redeem", async () => {
         const { vault, ousd, usdt, usdc, dai, anna, OUSDmetaStrategy } =
-          await loadFixture(withBalancedOUSDMetaPool);
+          await loadFixture(withCRV3TitledOUSDMetapool);
 
         await vault.connect(anna).allocate();
 
@@ -62,7 +62,7 @@ forkOnlyDescribe(
         // Total supply should be up by at least (10k x 2) + (10k x 2) + 10k = 50k
         const currentSupply = await ousd.totalSupply();
         const supplyAdded = currentSupply.sub(supplyBeforeMint);
-        expect(supplyAdded).to.be.gte(ousdUnits("50000"));
+        expect(supplyAdded).to.be.gte(ousdUnits("49999"));
 
         const currentBalance = await ousd.connect(anna).balanceOf(anna.address);
 
@@ -112,13 +112,16 @@ async function mintTest(fixture, user, asset, amount = "30000") {
   // Supply checks
   const newSupply = await ousd.totalSupply();
   const supplyDiff = newSupply.sub(currentSupply);
+  const ousdUnitAmount = ousdUnits(amount);
+
+  // The pool is titled to 3CRV by a million
   if ([usdt.address, usdc.address].includes(asset.address)) {
-    // Ensure 2x OUSD has been added to supply
+    // It should have added amount*3 supply
     // (in case of USDT/USDC)
-    expect(supplyDiff).to.approxEqualTolerance(ousdUnits(amount).mul(2), 1);
+    expect(supplyDiff).to.approxEqualTolerance(ousdUnitAmount.mul(3), 5);
   } else {
     // 1x for DAI
-    expect(supplyDiff).to.approxEqualTolerance(ousdUnits(amount), 2);
+    expect(supplyDiff).to.approxEqualTolerance(ousdUnitAmount, 1);
   }
 
   // Ensure some LP tokens got staked under OUSDMetaStrategy address
@@ -133,9 +136,6 @@ async function mintTest(fixture, user, asset, amount = "30000") {
     expect(rewardPoolBalanceDiff).to.equal("0");
   } else {
     // Should have staked the LP tokens for USDT and USDC
-    expect(rewardPoolBalanceDiff).to.approxEqualTolerance(
-      ousdUnits(amount).mul(2),
-      5
-    );
+    expect(rewardPoolBalanceDiff).to.be.gte(ousdUnits(amount).mul(3).div(2));
   }
 }
